% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{add_params_list}
\alias{add_params_list}
\title{Add parameters to a tick_transitions tibble}
\usage{
add_params_list(tick_transitions, parameters = tick_params)
}
\arguments{
\item{tick_transitions}{Tick transitions tibble}

\item{parameters}{Tick parameters tibble}
}
\value{
A new tick_transitions tibble with a column, params_list
that is a named list of the parameters for each transition.
}
\description{
Add parameters to a tick_transitions tibble
}
