% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{age}
\alias{age}
\title{Get tick age based on life stage}
\usage{
age(life_stage)
}
\arguments{
\item{life_stage}{Three-character string representing tick life stage}
}
\value{
Character representing tick age, where "l" indicates larvae, "n"
indicates nymph, and "a" indicates adult
}
\description{
Get tick age based on life stage
}
\examples{
age("f_l")
}
