% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{annual_growth_rate}
\alias{annual_growth_rate}
\title{Calculate annual growth rate}
\usage{
annual_growth_rate(out)
}
\arguments{
\item{out}{Model output data frame}
}
\value{
Numeric vector of length one representing the annual factor by
which the total tick population changes. To use this function, it is best
to run the model for at least three years.
}
\description{
Calculate annual growth rate
}
\examples{
\dontrun{
out <- run(ogden2005)
annual_growth_rate(out)
}

}
