% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_configs.R
\docType{data}
\name{config_ex_1}
\alias{config_ex_1}
\title{Simple model configuration example}
\format{
A \code{\link{config}}
}
\usage{
config_ex_1
}
\description{
This model configuration uses only non-delay transitions, and no transitions
depend on predictors (e.g. weather or host community). Parameter values are
selected so that the population is stable over time.
}
\keyword{datasets}
