% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_configs.R
\docType{data}
\name{config_ex_2}
\alias{config_ex_2}
\title{Simple model configuration example using delays}
\format{
A \code{\link{config}}
}
\usage{
config_ex_2
}
\description{
This model configuration uses delay transitions for all transitions except
the adult to eggs transition. As in \code{config_ex_1}, no transitions depend
on predictors, and the population is stable over time.
}
\examples{
TODO

}
\keyword{datasets}
