% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_transition_functions.R
\name{constant_fun}
\alias{constant_fun}
\title{Constant function}
\usage{
constant_fun(a)
}
\arguments{
\item{a}{Parameter \code{a} in parameters table.}
}
\value{
Numeric vector of length 1 equal to input parameter \code{a}
}
\description{
Constant function
}
\examples{
constant_fun(NULL, NULL, 1)
}
