% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{empty_delay_array}
\alias{empty_delay_array}
\title{Generate an empty delay array
Dimensions are: from, to, time}
\usage{
empty_delay_array(life_stages, steps, max_duration)
}
\description{
Generate an empty delay array
Dimensions are: from, to, time
}
