% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{empty_population_matrix}
\alias{empty_population_matrix}
\title{Create an empty (zero population) population matrix}
\usage{
empty_population_matrix(life_stages, steps)
}
\description{
Create an empty (zero population) population matrix
}
