% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{empty_transition_matrix}
\alias{empty_transition_matrix}
\title{Create an empty matrix of transition probabilities between life stages}
\usage{
empty_transition_matrix(life_stages)
}
\description{
Create an empty matrix of transition probabilities between life stages
}
