% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_functions.R
\name{expo_fun}
\alias{expo_fun}
\title{Exponential function}
\usage{
expo_fun(x, y, a, b)
}
\arguments{
\item{x}{Predictor 1 in transitions table.}

\item{y}{Predictor 2 in transitions table. Not used in this function.}

\item{a}{Parameter \code{a} in parameters table.}

\item{b}{Parameter \code{b} in parameters table.}
}
\value{
Numeric vector of length 1
}
\description{
Exponential function
}
\examples{
expo_fun(.5, NULL, .1, .3)

}
