% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_specification.R
\name{format.predictor_spec}
\alias{format.predictor_spec}
\title{Format a \code{predictor_spec}}
\usage{
\method{format}{predictor_spec}(x, ...)
}
\arguments{
\item{x}{a predictor_spec}

\item{...}{not used}
}
\value{
string representation of input
}
\description{
Format a \code{predictor_spec}
}
