% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_specification.R
\name{format.predictors_spec_node}
\alias{format.predictors_spec_node}
\title{Format a \code{predictor_spec_node}}
\usage{
\method{format}{predictors_spec_node}(x, ...)
}
\arguments{
\item{x}{a predictors_spec_node}

\item{...}{not used}
}
\value{
string representation of input
}
\description{
Format a \code{predictor_spec_node}
}
