% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{gen_trans_matrix}
\alias{gen_trans_matrix}
\title{Generate a matrix of transition probabilities between tick life stages}
\usage{
gen_trans_matrix(time, N, N_developing)
}
\arguments{
\item{time}{Numeric vector indicating day to get transition probabilities}

\item{N}{Tick population matrix. See get_tick_den for details.}

\item{N_developing}{Matrix of currently developing ticks. See get_tick_den
for details.

TODO pass life_stages, tick_transitions, any other global vars...}
}
\value{
Matrix of transition probabilities, indicating the probabilities of
transitioning from each stage (axis 1) to each stage (axis 2).
}
\description{
Generate a matrix of transition probabilities between tick life stages
}
\details{
TODO
}
