% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{get_host_den}
\alias{get_host_den}
\title{TODO need to pass host_comm as param wherever function is called}
\usage{
get_host_den(time)
}
\description{
TODO need to pass host_comm as param wherever function is called
}
