% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{get_params}
\alias{get_params}
\title{Return list of parameters for a given transition}
\usage{
get_params(from, to, parameters)
}
\arguments{
\item{from}{String indicating life stage a transition is from}

\item{to}{String indicating life stage a transition is to}

\item{parameters}{Tick parameters tibble}
}
\value{
Named list of parameters needed for the transition function from the
\code{from} life stage to the \code{to} life stage.
}
\description{
Return list of parameters for a given transition
}
\details{
This function is used for an optimization where we preprocess the model
inputs by joining the transitions and parameters into a tibble once rather
than searching for parameters for each function each time the function is
called.
}
