% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{get_pred}
\alias{get_pred}
\title{Get the value of a predictor}
\usage{
get_pred(time, pred, is_delay, N, N_developing)
}
\arguments{
\item{time}{Numeric vector indicating span of days to get predictor values}

\item{pred}{String indicating which predictor, one of: temp, vpd, host_den or
NA}

\item{is_delay}{Boolean indicating whether the predictor is for a transition
involving a delay}
}
\description{
Get the value of a predictor
}
