% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{get_temp}
\alias{get_temp}
\title{Get temperature from input data}
\usage{
get_temp(time, weather)
}
\arguments{
\item{time}{Numeric vector indicating days to get temperature data}

\item{weather}{Data frame with a Julian date "j_day" column and average
temperature "tmean" column}
}
\value{
Numeric vector of temperature for the given period
}
\description{
Get temperature from input data
}
