% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{get_transition_val}
\alias{get_transition_val}
\title{Get the value determining probability or duration of a transition}
\usage{
get_transition_val(time, transition_row_with_parameters, N, N_developing)
}
\arguments{
\item{time}{Numeric vector indicating span of days to get predictor values
TODO rename param to be shorter. I don't think _with_parameters is needed}

\item{transition_row_with_parameters}{A row from the tick_transitions tibble
with parameters added.}

\item{N}{Tick population matrix. See get_tick_den for details.}

\item{N_developing}{Matrix of currently developing ticks. See get_tick_den
for details.}
}
\value{
Numeric vector indicating probability or duration of a transition.
Vector will be of length one if the transition is not a delay transition???
TODO better describe what this returns
}
\description{
Get the value determining probability or duration of a transition
}
\details{
This generic function pulls out the functional form and parameters needed to
make the transition function, then evaluates it using supplied predictors
(pred1, pred2) like temperature or host density. To identify a transition,
this approach takes an entire transition_row, since there can be multiple
rows with the same 'from' and 'to'.
}
