% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{get_vpd}
\alias{get_vpd}
\title{TODO need to pass weather as param wherever function is called}
\usage{
get_vpd(time)
}
\description{
TODO need to pass weather as param wherever function is called
}
