% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{graph_lifecycle}
\alias{graph_lifecycle}
\title{Visualize transitions as a life cycle graph}
\usage{
graph_lifecycle(transitions)
}
\arguments{
\item{transitions}{Tick transitions tibble}
}
\description{
Visualize transitions as a life cycle graph
}
\details{
This function could be used to visually confirm that a custom \code{config} has
all the transitions intended
}
\examples{
graph_lifecycle(config_ex_1$transitions)
}
