% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{graph_population_each_group}
\alias{graph_population_each_group}
\title{Graph population size of each life stage over time}
\usage{
graph_population_each_group(output)
}
\arguments{
\item{output}{Model output; a tibble}
}
\value{
ggplot object
}
\description{
Graph population size of each life stage over time
}
\examples{
out <- run(config_ex_1)
graph_population_each_group(out)

}
