% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_configs.R
\docType{data}
\name{host_example_config}
\alias{host_example_config}
\title{Configuration for showing how we can modify host community data}
\format{
A \code{\link{config}}
}
\usage{
host_example_config
}
\description{
Configuration for showing how we can modify host community data
}
\keyword{datasets}
