% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_configs.R
\docType{data}
\name{infect_example_config}
\alias{infect_example_config}
\title{Configuration for showing infection dynamics}
\format{
A \code{\link{config}}
}
\usage{
infect_example_config
}
\description{
Configuration for showing infection dynamics
}
\keyword{datasets}
