% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{infected}
\alias{infected}
\title{Get tick infection status}
\usage{
infected(life_stage)
}
\arguments{
\item{life_stage}{Three-character string representing tick life stage}
}
\value{
Boolean indicating whether a life_stage is infected
}
\description{
Get tick infection status
}
\examples{
infected("fin")
}
