% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_cycle.R
\name{life_cycle}
\alias{life_cycle}
\title{Create a \code{life_cycle} from a collection of \code{transition}s}
\usage{
life_cycle(...)
}
\arguments{
\item{...}{A set of \code{\link{transition}}s}
}
\value{
a \code{life_cycle}
}
\description{
Create a \code{life_cycle} from a collection of \code{transition}s
}
