% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_function.R
\name{new_transition_function}
\alias{new_transition_function}
\title{Constructor for transition functions}
\usage{
new_transition_function(fun)
}
\description{
Transition functions must return a numeric vector. See
\code{\link{constant_fun}}, \code{\link{expo_fun}} and
\code{\link{infect_fun}} for examples for how to write custom functions.
}
