% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{output_to_df}
\alias{output_to_df}
\title{Convert model output to a nicely formatted tibble}
\usage{
output_to_df(out)
}
\arguments{
\item{out}{Output from run()}
}
\value{
Tibble with number of ticks per life stage per day.
}
\description{
Convert model output to a nicely formatted tibble
}
