% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Create a set of parameters}
\usage{
parameters(...)
}
\arguments{
\item{...}{A set of named numeric vectors, each corresponding to a parameter.
If a parameter is of length > 1, each element must be named.}
}
\value{
a \code{parameters} object
}
\description{
Create a set of parameters
}
\examples{
# create a set of scalar parameters
parameters(a = 1, b = 2)

# parameters of length > 1 may be useful for host-related parameters that
# differ between host species, for example tick feeding success
parameters(a = 1, feeding_success = c(deer = 0.49, squirrel = 0.17))
}
