% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_specification.R
\name{predictors_spec}
\alias{predictors_spec}
\title{Specify what predictors to use in a transition and how to use that data}
\usage{
predictors_spec(...)
}
\arguments{
\item{...}{named sequence of \code{predictor_specification_node}s. The names are
matched with the formal args to the containing \code{\link{transition}}'s
\code{fun} to determine which input in \code{fun} each predictor will be passed to.}
}
\value{
a \code{predictors_spec} object
}
\description{
Specify what predictors to use in a transition and how to use that data
}
