% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_cycle.R
\name{print.life_cycle}
\alias{print.life_cycle}
\title{Print a life cycle}
\usage{
\method{print}{life_cycle}(x, ..., max = 10L)
}
\arguments{
\item{x}{A \code{life_cycle}}

\item{max}{number of transitions to print, or NULL to print all transitions}
}
\description{
Print a life cycle
}
