% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_specification.R
\name{print.predictor_spec}
\alias{print.predictor_spec}
\title{Print a \code{predictor_spec}}
\usage{
\method{print}{predictor_spec}(x, ...)
}
\arguments{
\item{x}{a predictor_spec}

\item{...}{not used}
}
\description{
Print a \code{predictor_spec}
}
