% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_specification.R
\name{print.predictors_spec_node}
\alias{print.predictors_spec_node}
\title{Print a \code{predictors_spec_node}}
\usage{
\method{print}{predictors_spec_node}(x, ...)
}
\arguments{
\item{x}{a predictors_spec_node}

\item{...}{not used}
}
\description{
Print a \code{predictors_spec_node}
}
