% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{print.transition}
\alias{print.transition}
\title{Print a transition}
\usage{
\method{print}{transition}(x, ...)
}
\arguments{
\item{x}{A \code{transition}}

\item{...}{not used}
}
\description{
Print a transition
}
