% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{print_all_params}
\alias{print_all_params}
\title{Print parameter names and values for each transition}
\usage{
print_all_params(tick_transitions, parameters)
}
\arguments{
\item{tick_transitions}{Tick transitions table}

\item{parameters}{Tick parameters table}
}
\description{
Print parameter names and values for each transition
}
