% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{print_params}
\alias{print_params}
\title{Print parameters for a given transition}
\usage{
print_params(transition_row, parameters)
}
\arguments{
\item{transition_row}{A row from the tick transitions tibble.}

\item{parameters}{Tick parameters tibble.}
}
\description{
Print parameters for a given transition
}
\details{
Print the parameters that are being grabbed through pattern matching and the
name of the function. Useful for debugging to determine whether correct
parameters are being selected for each transition.
}
