% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{process}
\alias{process}
\title{Get tick current process}
\usage{
process(life_stage)
}
\arguments{
\item{life_stage}{Three-character string representing tick life stage}
}
\value{
Character representing current process tick is undergoing, where "a"
indicates attached, "e" indicates engorged, "h" indicates hardening, "q"
indicates questing, "r" indicates reproductive", and "f" indicates feeding
}
\description{
Get tick current process
}
\examples{
process("f_l")
}
