% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\title{create a config object from a YAML file}
\usage{
read_config(file)
}
\arguments{
\item{file}{YAML file to read}
}
\value{
A \code{config} object
}
\description{
create a config object from a YAML file
}
\examples{
\dontrun{
read_config('cfg.yml')
}

}
