% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{run}
\alias{run}
\title{Run the model}
\usage{
run(steps, initial_population)
}
\arguments{
\item{steps}{Numeric vector of length one indicating the duration to run the
model over in days.}

\item{initial_population}{Named???? numeric vector indicating the starting
population for each life stage. Length should be equal to the number of
life stages.}
}
\value{
List with population matrix, matrix of currently developing ticks,
and delay matrix.
TODO we probably want exported function to just return the nicely formatted
tibble of population per life stage over time.
}
\description{
Run the model
}
