% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{run_all_configs}
\alias{run_all_configs}
\title{Run the model for each config}
\usage{
run_all_configs(configs, parallel = FALSE)
}
\arguments{
\item{configs}{List of \code{config} objects}

\item{parallel}{Logical; if TRUE, run on all cores using \code{parallel} package.}
}
\value{
A stacked data frame of the model outputs for each \code{config}. Return
value is like \code{run()}, with an additional column "config" identifying
the \code{config} object that results were generated from.
}
\description{
Run the model for each config
}
