% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{set_initial_population}
\alias{set_initial_population}
\title{Set initial population for each life stage - zero if not specified in cfg}
\usage{
set_initial_population(population, initial_population)
}
\description{
Set initial population for each life stage - zero if not specified in cfg
}
