% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_functions.R
\name{snow_cover_fun}
\alias{snow_cover_fun}
\title{Mortality as a function of whether there is a snow on the ground}
\usage{
snow_cover_fun(x, y, no_snow_mort, snow_mort)
}
\arguments{
\item{x}{amount of snow on ground}

\item{y}{not used in this transition function}

\item{no_snow_mort}{mortality with no snow on the ground}

\item{snow_mort}{mortality with snow on the ground}
}
\description{
Mortality as a function of whether there is a snow on the ground
}
