% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_configs.R
\docType{data}
\name{temp_example_config}
\alias{temp_example_config}
\title{Configuration for showing how we can modify climate data}
\format{
A \code{\link{config}}
}
\usage{
temp_example_config
}
\description{
Configuration for showing how we can modify climate data
}
\keyword{datasets}
