% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{test_lifecycles}
\alias{test_lifecycles}
\title{Check that transitions form a cycle}
\usage{
test_lifecycles(life_stages, tick_transitions, graph = TRUE)
}
\arguments{
\item{life_stages}{Character vector of life stages.}

\item{tick_transitions}{Tick transitions tibble}

\item{graph}{Default \code{TRUE}, if \code{FALSE}, graph is not created.}
}
\description{
Check that transitions form a cycle
}
\details{
Tests to ensure that there are no "dead-ends" in the life cycle by ensuring
that all non-mortality life stages occur in both the \code{from} and \code{to} fields
of transitions.

Optionally, also outputs a graph of the life cycle for visually confirming
that transitions are as intended.
}
