% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{test_transitions}
\alias{test_transitions}
\title{Calculate transition value for each transition}
\usage{
test_transitions(tick_transitions, tick_params, max_delay, host_comm, weather)
}
\arguments{
\item{tick_transitions}{Tick transitions tibble}

\item{tick_params}{Tick parameters tibble}

\item{max_delay}{Numeric vector of length one. Determines the maximum
number of days that a delayed transition can last.}

\item{host_comm}{Host community tibble.}

\item{weather}{Weather tibble.}
}
\description{
Calculate transition value for each transition
}
\details{
For use in testing. Much faster than running the whole model and useful for
seeing which functions are being problematic when the model breaks
}
