% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_function.R
\name{transition_function_is_custom}
\alias{transition_function_is_custom}
\title{Return whether a transition function is custom or defined in the package}
\usage{
transition_function_is_custom(fun)
}
\description{
Return whether a transition function is custom or defined in the package
}
