% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{update_delay_arr}
\alias{update_delay_arr}
\title{Update the delay array for the current time}
\usage{
update_delay_arr(time, delay_arr, N, N_developing)
}
\arguments{
\item{time}{Numeric vector indicating day to get transition probabilities}

\item{delay_arr}{Delay array from previous time step}

\item{N}{Tick population matrix. See get_tick_den for details.}

\item{N_developing}{Matrix of currently developing ticks. See get_tick_den
for details.}
}
\value{
Delay array indicating the number of ticks currently undergoing delay
transitions. Axis 1 is the from_stage, axis 2 is the to_stage, and axis 3
is the day on which the ticks will emerge from the transition. The value
at a given cell is the number of ticks emerging from the transition.
}
\description{
Update the delay array for the current time
}
\details{
TODO
}
