% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Create a \code{config} object}
\usage{
config(
  initial_population,
  transitions,
  parameters,
  host_comm,
  weather,
  steps,
  max_delay = 365L
)
}
\arguments{
\item{initial_population}{Named numeric vector indicating starting population
for each life stage. Life stages not specified are assumed to be 0.}

\item{transitions}{Tick transitions data frame. See readme for details.}

\item{parameters}{Tick parameters tibble. See readme for details.}

\item{host_comm}{Host community tibble. See readme for details.}

\item{weather}{Weather tibble. See readme for details.}

\item{steps}{Numeric vector of length one indicating the duration to run the
model over in days.}

\item{max_delay}{Numeric vector of length one. Determines the maximum
number of days that a delayed transition can last.}
}
\value{
A \code{config} object
}
\description{
Create a \code{config} object
}
