% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{get_pred}
\alias{get_pred}
\title{Get the value of a predictor}
\usage{
get_pred(
  time,
  pred,
  is_delay,
  N,
  N_developing,
  max_delay,
  life_stages,
  host_comm,
  weather
)
}
\arguments{
\item{time}{Numeric vector indicating span of days to get predictor values}

\item{pred}{String indicating which predictor, one of: temp, vpd, host_den or
NA}

\item{is_delay}{Boolean indicating whether the predictor is for a transition
involving a delay}

\item{N}{Tick population matrix. See get_tick_den for details.}

\item{N_developing}{Matrix of currently developing ticks. See get_tick_den
for details.}

\item{max_delay}{Numeric vector of length one. Determines the maximum
number of days that a delayed transition can last.}

\item{life_stages}{Character vector of life stages.}

\item{host_comm}{Host community tibble.}

\item{weather}{Weather tibble.}
}
\description{
Get the value of a predictor
}
