% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{get_pred_from_table}
\alias{get_pred_from_table}
\title{Get a predictor from input data}
\usage{
get_pred_from_table(time, table, col)
}
\arguments{
\item{time}{Numeric vector of days to get data. Ignored if input is constant
over time (as indicated by absence of 'j_day' column)}

\item{table}{input weather or host density data frame from config object}

\item{col}{name of column from table to get predictor values from}
}
\description{
Get a predictor from input data
}
