% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{get_pred_from_table}
\alias{get_pred_from_table}
\title{Get a predictor from input data}
\usage{
get_pred_from_table(time, pred, table)
}
\arguments{
\item{time}{Numeric vector of days to get data. Ignored if input is constant
over time (as indicated by NA value in 'j_day' column)}

\item{pred}{string specifying the name of the predictor, e.g. "host_den"}

\item{table}{input predictors table}
}
\description{
Get a predictor from input data
}
