% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{get_tick_den}
\alias{get_tick_den}
\title{Get tick density for specified time and life stages}
\usage{
get_tick_den(time, N, N_developing, pred, life_stages)
}
\arguments{
\item{time}{Numeric vector of length one indicating day to get tick density}

\item{N}{Matrix of number of ticks not currently undergoing development
per life stage per day}

\item{N_developing}{Matrix of number of currently developing ticks per life
stage per day}

\item{pred}{Character vector indicating life stages to get tick
density of}

\item{life_stages}{Character vector of life stages.}
}
\value{
Numeric vector of length one indicating current number of ticks in
given life stages
}
\description{
Get tick density for specified time and life stages
}
