% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{graph_population_each_group}
\alias{graph_population_each_group}
\title{Graph population size of each life stage over time}
\usage{
graph_population_each_group(out_N_df, title = NULL)
}
\arguments{
\item{out_N_df}{Tibble of population size per life stage per day}

\item{title}{Optional title to display on graph}
}
\value{
ggplot object
}
\description{
Graph population size of each life stage over time
}
