% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{graph_population_overall_trend}
\alias{graph_population_overall_trend}
\title{Graph overall trend in population}
\usage{
graph_population_overall_trend(output)
}
\arguments{
\item{output}{Model output; a tibble}
}
\value{
ggplot object
}
\description{
Graph overall trend in population
}
\details{
See roughly whether population is increasing or decreasing.
}
