% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{run}
\alias{run}
\title{Run the model}
\usage{
run(
  steps,
  initial_population,
  tick_transitions,
  tick_params,
  life_stages,
  max_delay,
  host_comm,
  weather
)
}
\arguments{
\item{steps}{Numeric vector of length one indicating the duration to run the
model over in days.}

\item{initial_population}{Named???? numeric vector indicating the starting
population for each life stage. Length should be equal to the number of
life stages.}

\item{tick_transitions}{Tick transitions tibble}

\item{tick_params}{Tick parameters tibble}

\item{life_stages}{Character vector of life stages.}

\item{max_delay}{Numeric vector of length one. Determines the maximum
number of days that a delayed transition can last.}

\item{host_comm}{Host community tibble.}

\item{weather}{Weather tibble.}
}
\value{
List with population matrix, matrix of currently developing ticks,
and delay matrix.
TODO we probably want exported function to just return the nicely formatted
tibble of population per life stage over time.
}
\description{
Run the model
}
