% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{run}
\alias{run}
\title{Run the model}
\usage{
run(cfg, progress = TRUE)
}
\arguments{
\item{cfg}{An \code{IxPopDyMod::config} object}

\item{progress}{Boolean indicating whether to log progress every 100 steps}
}
\value{
Data frame of population of ticks of each life stage each day
}
\description{
Run the model
}
\examples{
run(config_ex_1)

}
