% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{run}
\alias{run}
\title{Run the model}
\usage{
run(cfg)
}
\arguments{
\item{cfg}{An \code{IxPopDyMod::config} object}
}
\value{
Data frame of population of ticks of each life stage each day
}
\description{
Run the model
}
