% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{run_all_configs}
\alias{run_all_configs}
\title{Run the model for each config}
\usage{
run_all_configs(configs, parallel = FALSE)
}
\arguments{
\item{configs}{List of \code{config} objects}

\item{parallel}{Logical; if TRUE, run on all cores using \code{parallel} package.}
}
\value{
A list of data frame model outputs like those returned by \code{run()}
}
\description{
Simple convenience wrapper for calling run on each \code{config} in a list
}
\examples{
# run two example configs and save results

\dontrun{
outputs <- run_all_configs(list(config_ex_1, config_ex_2))
}
}
