% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{transition}
\alias{transition}
\title{Create a \code{transition} object}
\usage{
transition(
  from,
  to,
  fun,
  transition_type,
  mortality_type = NULL,
  predictors = NULL,
  parameters = list()
)
}
\arguments{
\item{from}{The name of the life stage a tick is transitioning from.}

\item{to}{The name of the life stage a tick is transitioning to, or NULL if
the transition is representing mortality.}

\item{fun}{The \code{\link{transition_function}} to evaluate.}

\item{transition_type}{One of:
\code{"probability"}: the evaluated transition is interpreted as the daily
fraction of ticks that complete the transition. Ticks remain in the
original life stage if they do not complete a transition or undergo
mortality.
\code{"duration"}: the transition is complete on the first day that the
cumulative sum of the evaluated transition is greater than or equal to 1.
No ticks remain in the original life stage at the end of a transition -
they either complete the transition or die.}

\item{mortality_type}{One of:
\code{NULL}: the default, indicating that the transition is not mortality.
\code{"per_day"}: indicates that the evaluated transition is the fraction of
ticks that dies each day.
\code{"throughout_transition"}: only valid for \code{"duration"} type transitions,
where it indicates that the evaluated transition is the fraction of
ticks that die throughout the entire transition.}

\item{predictors}{Optional, named character vector of predictors to use in
evaluating \code{fun}. Names are matched with the formal args to \code{fun} to
determine which input in \code{fun} each predictor will be passed to. Each value
can be one of:
- A string in the \code{"pred"} column in the \code{\link{predictors}} table.
In this case, the predictor value passed to \code{fun} is the corresponding
value of that predictor in the table.
- A string that matches at least one life stage name via regex. In this
case, the value passed to \code{fun} is the sum of the population sizes of
all matched life stages.}

\item{parameters}{Optional, a \code{\link{parameters}} object, or a named
list of numeric vectors.

TODO need to actually create the helper functions linked in these docs.}
}
\value{
a \code{transition} object
}
\description{
A \code{transition} object represents a single transition between two tick life
stages, or the mortality rate from a life stage.
}
