% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{vary_many_params}
\alias{vary_many_params}
\title{Generate copies of a \code{config} with all combinations of modified parameters}
\usage{
vary_many_params(cfg, param_rows, values_list)
}
\arguments{
\item{cfg}{Base \code{config} to make modified copies of}

\item{param_rows}{Numeric vector indicating the rows in the parameters
table where parameter values should be modified. Length must equal length
of values_list}

\item{values_list}{List of numeric vectors. The values of a vector
\code{values_list[[i]]} are the parameter values to use for the parameter
identified by \code{param_rows[[i]]}}
}
\value{
A list of \code{config}s
}
\description{
Generate copies of a \code{config} with all combinations of modified parameters
}
