% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{vary_param}
\alias{vary_param}
\title{Generate copies of a config with a modified parameter
TODO locate parameter by row in parameters table or filtering
with from, to, param_name, and host_spp?
TODO Dave}
\usage{
vary_param(config, param_row, values)
}
\arguments{
\item{config}{base configuration to make modified copies of}

\item{param_row}{row number of parameter to vary}

\item{values}{Numeric vector of values to use for parameter}
}
\value{
List of configs?????
}
\description{
Generate copies of a config with a modified parameter
TODO locate parameter by row in parameters table or filtering
with from, to, param_name, and host_spp?
TODO Dave
}
